Add user can only delete own account

+++ plantplanner/routes.py
@@ -148,27 +148,32 @@ def dashboard():
 
 
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully!")
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template(
-            "add_user.html",
-            form=form,
-            name=name,
-            our_users=our_users)
-    except:
-        flash("Sorry, there was a problem deleting the user")
-        return render_template(
-            "add_user.html",
-            form=form,
-            name=name,
-            our_users=our_users)
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully!")
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template(
+                "add_user.html",
+                form=form,
+                name=name,
+                our_users=our_users)
+        except:
+            flash("Sorry, there was a problem deleting the user")
+            return render_template(
+                "add_user.html",
+                form=form,
+                name=name,
+                our_users=our_users)
+    else:
+        flash("Sorry, you can't delete that user")
+        return redirect(url_for('dashboard'))
 
 
 @app.route('/posts/delete/<int:id>')
